// src/interfaces/order.interface.ts
// Generated by https://quicktype.io

export interface Order {
    id:                    string;
    grand_total:           number;
    total_items:           number;
    buyer_id:              string;
    payment_type:          string;
    payment_id:            string;
    payment_status:        number;
    payment_url:           string;
    shipping_address_1:    string;
    shipping_address_2:    null;
    shipping_city:         string;
    shipping_state:        string;
    shipping_zip:          number;
    shipping_address_name: string;
    created_at:            string;
    updated_at:            string;
    deleted_at:            null;
    buyer:                 Buyer;
    order_items:           OrderItem[];
}

export interface Buyer {
    id:         string;
    first_name: string;
    last_name:  string;
    email:      string;
    user_name:  string;
}

export interface OrderItem {
    id:               string;
    order_id:         string;
    product_id:       string;
    product_quantity: number;
    product_price:    number;
    created_at:       string;
    updated_at:       string;
    deleted_at:       null;
    product:          Product;
}

export interface Product {
    id:   string;
    name: string;
}